refactor: Remove prop initial option.
BREAKING CHANGE: remove prop initial option #991.
