--- src/util/get-default-value.js
@@ -1,5 +1,0 @@
-export default function getDefaultValue (elem, propDef) {
-  return typeof propDef.default === 'function'
-    ? propDef.default(elem, { name: propDef.nameOrSymbol })
-    : propDef.default;
-}
