--- src/util/get-initial-value.js
@@ -1,5 +1,0 @@
-export default function getInitialValue (elem, propDef) {
-  return typeof propDef.initial === 'function'
-    ? propDef.initial(elem, { name: propDef.nameOrSymbol })
-    : propDef.initial;
-}
