--- src/util/get-prop-data.js
@@ -1,6 +1,0 @@
-import data from './data';
-
-export default function getPropData (elem, name) {
-  const elemData = data(elem, 'props');
-  return elemData[name] || (elemData[name] = {});
-}
